refactor(app): rename run entity directory to pipeline-run
The directory name 'run' is ambiguous — 'pipeline-run' matches the
resource it represents and aligns with the rpc resource-types naming.
Entity ID and URL segments are unchanged.
