+++ javascript/app/config/entities/pipeline-run/detail.ts
