+++ javascript/app/config/entities/pipeline/detail.ts
@@ -1,6 +1,6 @@
 import { CellType } from '@michelangelo-ai/core';
 
-import { SHARED_RUN_CELL_CONFIG } from '../run/shared';
+import { SHARED_RUN_CELL_CONFIG } from '../pipeline-run/shared';
 import { PIPELINE_STATE_CELL, PIPELINE_TYPE_CELL } from './shared';
 
 import type { DetailViewConfig } from '@michelangelo-ai/core';
