+++ javascript/app/config/phases/data.ts
@@ -1,5 +1,5 @@
 import { PIPELINE_ENTITY_CONFIG } from '../entities/pipeline/pipeline';
-import { RUN_ENTITY_CONFIG } from '../entities/run/run';
+import { PIPELINE_RUN_ENTITY_CONFIG } from '../entities/pipeline-run/pipeline-run';
 
 import type { PhaseConfig } from '@michelangelo-ai/core';
 
@@ -12,7 +12,7 @@ export const DATA_PHASE: PhaseConfig = {
   state: 'disabled' as const,
   entities: [
     { ...PIPELINE_ENTITY_CONFIG, state: 'disabled' },
-    { ...RUN_ENTITY_CONFIG, state: 'disabled' },
+    { ...PIPELINE_RUN_ENTITY_CONFIG, state: 'disabled' },
     {
       id: 'datasources',
       name: 'data sources',
